refactor: [FE] Controller 리팩토링
- Controller에서 Command 관련 함수 CommandController로 이동
